refactor. shared ui 컴포넌트 분리
