+++ src/pages/PostsManagerPage.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react"
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { useEffect, useState } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 import {
   Button,
@@ -7,16 +7,13 @@ import {
   CardContent,
   CardHeader,
   CardTitle,
-  Dialog,
   DialogContent,
   DialogHeader,
   DialogTitle,
   Input,
-  Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
-  SelectValue,
   Table,
   TableBody,
   TableCell,
@@ -25,6 +22,8 @@ import {
   TableRow,
   Textarea,
 } from "../shared/ui"
+import { Dialog } from "../shared/ui/Dialog"
+import { Select, SelectValue } from "../shared/ui/Select"
 
 const PostsManager = () => {
   const navigate = useNavigate()
@@ -268,7 +267,6 @@ const PostsManager = () => {
   // 댓글 좋아요
   const likeComment = async (id, postId) => {
     try {
-
       const response = await fetch(`/api/comments/${id}`, {
         method: "PATCH",
         headers: { "Content-Type": "application/json" },
@@ -277,7 +275,9 @@ const PostsManager = () => {
       const data = await response.json()
       setComments((prev) => ({
         ...prev,
-        [postId]: prev[postId].map((comment) => (comment.id === data.id ? {...data, likes: comment.likes + 1} : comment)),
+        [postId]: prev[postId].map((comment) =>
+          comment.id === data.id ? { ...data, likes: comment.likes + 1 } : comment,
+        ),
       }))
     } catch (error) {
       console.error("댓글 좋아요 오류:", error)
