+++ src/shared/ui/Table.tsx
@@ -1,0 +1,37 @@
+export const TableHeader = ({ className, ...props }: React.HTMLAttributes<HTMLTableSectionElement>) => (
+  <thead className={`[&_tr]:border-b ${className}`} {...props} />
+)
+TableHeader.displayName = "TableHeader"
+
+export const TableHead = ({ className, ...props }: React.HTMLAttributes<HTMLTableCellElement>) => (
+  <th
+    className={`h-10 px-2 text-left align-middle font-medium text-muted-foreground [&:has([role=checkbox])]:pr-0 ${className}`}
+    {...props}
+  />
+)
+TableHead.displayName = "TableHead"
+
+export const TableCell = ({ className, ...props }: React.HTMLAttributes<HTMLTableCellElement>) => (
+  <td className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`} {...props} />
+)
+TableCell.displayName = "TableCell"
+
+export const TableBody = ({ className, ...props }: React.HTMLAttributes<HTMLTableSectionElement>) => (
+  <tbody className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+)
+TableBody.displayName = "TableBody"
+
+export const TableRow = ({ className, ...props }: React.HTMLAttributes<HTMLTableRowElement>) => (
+  <tr
+    className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+    {...props}
+  />
+)
+TableRow.displayName = "TableRow"
+
+export const Table = ({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) => (
+  <div className="w-full overflow-auto">
+    <table className={`table-fixed w-full caption-bottom text-sm ${className}`} {...props} />
+  </div>
+)
+Table.displayName = "Table"
